Moving unify test cases with subclasses into a separate function)
